[dist] Ensure `python3 -m build` works by default
Include pre-built documentation with distribution and improve error messaging
when Tkinter is not available. Simplify download page and recommend pipx.

This is now the tagged release for v1.6.2 however no changes to the Python
or Windows distributions has been made, the only difference is for the
source distribution here on Github and in the releases branch.