[app] Fix overlays stuck on #218
The logic to check bounding box is a bit complicated in the shared module
as it's different for the CLI vs. config file. This changes it to make the
differences more clear, and also updates the flag value from False to True.
This simplifies checks, and also allows us to fallback so that if the value
is not a boolean, we can assume the user specified smoothing time via the
`-bb` CLI flag (in which case the config option is overriden). When the `-bb`
flag is set alone, or from the UI, the bounding-box-smooth-time setting
is used as expected.